requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
pytesseract>=0.3.10
easyocr>=1.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
python-multipart>=0.0.6
pyyaml>=6.0
Pillow>=9.0.0
scipy>=1.9.0
scikit-learn>=1.1.0
